много всего, в том числе исправления странных ошибок

+++ src/apic.c
@@ -4,8 +4,8 @@
 
 
 void apic_interrupt(struct APIC* apic, char id) {
-	char cpu_id = apic->int_table[id];
+	char cpu_id = apic->int_table[(int)id]; // really, why gcc mark char id as warning?
 
-	if (apic->cpu->cores[cpu_id].state & INTERRUPTS)
-		core_int(&apic->cpu->cores[cpu_id], id);
+	if (apic->cpu->cores[(int)cpu_id].state & INTERRUPTS)
+		core_int(&apic->cpu->cores[(int)cpu_id], id);
 }
